autocraft now crafts blueprints first
Old behaviour: Autocraft used to craft parchment, manuscript, compendium and blueprints in order. this means that all the science would be used up crafting compendiums and not blueprints.

New behaviour: Now blueprints are crafted first and parchment, manuscript and compendium after. This means that science are used for blueprints first and for compendiums after.